requirements.txt
numpy>=1.14
Pillow>=5
pygame>=1.9
PyOpenGL>=3.1
PyGLM
